FYP/requirements.txt
fastapi
uvicorn
python-dotenv
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-google-genai==1.0.10
google-generativeai
# Removed sentence-transformers and huggingface to save space
faiss-cpu
pandas
pypdf